requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
pandas==2.0.3
flask==2.3.2
flask-socketio==5.3.4
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
scikit-fuzzy==0.4.2
paho-mqtt==1.6.1
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0